fix #199
